Adjusted Reverse DCC Support and Network Functions
Adjusted the Reverse DCC support so it should now work on all systems
(getting wan address, listening on all instead of a single local ip).

Also adjusted the support so that if the bot doesn't connect to the us
after 1 minute ( will probably decrease this) that we abort and stop
trying, the pack will then be marked as failed.

Network functions for determining IP addresses (WAN and LAN) as well as
IP manipulation were added to XG.Business.Helper under the class
NetworkActions.